# Reduce allocations in HashFile