--- a/lib/scanner/.gitignore
+++ b/lib/scanner/.gitignore
@@ -1,0 +1,1 @@
+_random.data