@AudriusButkevicius 2015-10-27
I wonder if a simillar change can be done in the request loop