@AudriusButkevicius 2015-10-27
Probably not, as that hashes a single block off the network.